requirements.txt
anyio==3.3.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
cycler==0.10.0
dash==1.20.0
dash-core-components==1.16.0
dash-cytoscape==0.3.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
DateTime==4.3
decorator==4.4.2
defusedxml==0.7.1
docker==5.0.0
english==2020.7.0
entrypoints==0.3
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
greenlet==1.1.2
h11==0.12.0
http-client==0.1.22
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.11.1
jupyterlab==3.2.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==1.6
nbclassic==0.3.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
neo4j==4.3.1
neotime==1.7.4
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.3
notebook==6.4.0
numpy==1.20.3
outcome==1.1.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
pansi==2020.7.3
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==4.14.3
plotly-express==0.4.1
prometheus-client==0.11.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py2neo==2021.1.5
pycparser==2.20
pycurl==7.44.1
Pygments==2.9.0
pymotif==0.0.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-pushsafer==1.0
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.9.30
requests==2.25.1
requests-unixsocket==0.2.0
retrying==1.3.3
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
selenium==4.1.0
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.25
statsmodels==0.13.1
terminado==0.10.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wsproto==1.0.0
zope.interface==5.4.0
